따로 직접 변경 안하고 변수로 보내기, 평균 실수까지 표시

+++ chapter07/src/ddit/chap07/sec02/Student.java
@@ -21,7 +21,7 @@ public class Student extends Person {
 	
 	public void calculateScore() {
 		sum = kor + eng + mat;
-		avg = (kor + eng + mat)/3;
+		avg = (kor + eng + mat)/(double)3;
 	}
 	
 	public void printScore() {
